refactor homepage validators
